Producto con base implementada

+++ Alimar/BL/articulo.cs
@@ -10,7 +10,7 @@ namespace BL
     {
         //permite crear un objeto de articulo
 
-        public articulo(string tipo, string path, string genero, int edad, string descipcion, int precio, string nombre)
+        public articulo(string tipo, string path, string genero, double edad, string descipcion, double precio, string nombre)
         {
             this.tipo = tipo;
             this.path = path;
@@ -25,9 +25,9 @@ namespace BL
         public string tipo { set; get; }
         public string path { set; get; }
         public string genero { set; get; }
-        public int edad { set; get; }
+        public double edad { set; get; }
         public string descipcion { set; get; }
-        public int precio { set; get; }
+        public double precio { set; get; }
         public string nombre { set; get; }
 
         public articulo()
